requirements.txt
# Synaptiq Backend Dependencies

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sse-starlette>=2.0.0

# Data Validation
pydantic>=2.6.0
pydantic-settings>=2.1.0
email-validator>=2.0.0

# Database - PostgreSQL
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0

# Authentication
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0

# Vector Store
qdrant-client>=1.7.0

# MongoDB (for legacy support)
motor>=3.3.0
pymongo>=4.6.0

# Task Queue
celery[redis]>=5.3.0
redis>=5.0.0

# HTTP Client
httpx>=0.26.0

# LLM / AI
openai>=1.12.0
tiktoken>=0.6.0
openai-agents>=0.1.0

# Data Extraction
supadata>=1.0.0

# CLI
click>=8.1.0
rich>=13.7.0

# Utilities
python-dotenv>=1.0.0
tenacity>=8.2.0
structlog>=24.1.0
